Moved coordinate entry component into MapFilter since getting 'Clear' to work properly as it was set up would require extremely awkward workarounds for a component that wouldn't ever be reused anyway (to fix when coordinates had been entered but not applied as a valid geometry, 'Clear' did not reset text boxes)
